+++ domain/src/main/java/woowacourse/movie/domain/seat/Grade.kt
@@ -1,6 +1,6 @@
 package woowacourse.movie.domain.seat
 
-import woowacourse.movie.domain.ticket.Price
+import woowacourse.movie.domain.price.Price
 
 enum class Grade(val price: Price) {
     B(Price(10000)), S(Price(15000)), A(Price(12000))
