+++ domain/src/test/java/woowacourse/movie/domain/MorningPolicyTest.kt
@@ -3,7 +3,7 @@ package woowacourse.movie.domain
 import junit.framework.TestCase.assertEquals
 import org.junit.Test
 import woowacourse.movie.domain.policy.MorningPolicy
-import woowacourse.movie.domain.ticket.Price
+import woowacourse.movie.domain.price.Price
 import java.time.LocalDate
 import java.time.LocalTime
