+++ domain/src/test/java/woowacourse/movie/domain/PriceCalculatorTest.kt
@@ -5,8 +5,8 @@ import org.junit.Test
 import woowacourse.movie.domain.policy.MorningPolicy
 import woowacourse.movie.domain.policy.MovieDayPolicy
 import woowacourse.movie.domain.policy.NightPolicy
-import woowacourse.movie.domain.ticket.Price
-import woowacourse.movie.domain.ticket.PriceCalculator
+import woowacourse.movie.domain.price.Price
+import woowacourse.movie.domain.price.PriceCalculator
 import java.time.LocalDateTime
 
 class PriceCalculatorTest {
